% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{average_model_SE}
\alias{average_model_SE}
\title{Compute the average model SE}
\usage{
average_model_SE(SE)
}
\arguments{
\item{SE}{the SD of the posterior distribution for each parameter in the simulation scenario.}
}
\value{
average model SE
}
\description{
Compute the average model SE
}
\details{
Note that the SE here is not computed around the point estimate. Rather, it is taken from the posterior distribution of the parameter. We input SE (standard error), the formula needs the estimated variance. Hence, the SE is squared in the formula below.
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
