% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{average_model_SE_mcmc_se}
\alias{average_model_SE_mcmc_se}
\title{Compute the MCMC SE for the average model SE}
\usage{
average_model_SE_mcmc_se(SE, n_sim, modSE)
}
\arguments{
\item{SE}{the SD of the posterior distribution for each parameter in the simulation scenario.}

\item{nsim}{scalar. Number of iterations used in the scenario.}
}
\value{
average model SE MCMC SE
}
\description{
Compute the MCMC SE for the average model SE
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
