% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jasper.R
\name{burn}
\alias{burn}
\title{Burn function for model output}
\usage{
burn(x, ...)
}
\arguments{
\item{x}{an mHMM object}
}
\value{
mHMM_cont object for which the burn-in samples have been removed
        for each parameter.
}
\description{
Burn function for model output
}
