% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{coverage}
\alias{coverage}
\title{Compute coverage}
\usage{
coverage(CI, true_param_value)
}
\arguments{
\item{CI}{list. Each element contains the upper and lower values of the 95\% CI of an interation.}

\item{true_param_value}{numeric scalar. True value of the parameter.}
}
\value{
numeric vector with two values: (1) Probability that the 95\% CI contains the true value and (2) MCMC Standard Error
}
\description{
Compute coverage
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
