% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{coverage_MCMC_SE}
\alias{coverage_MCMC_SE}
\title{Coverage MCMC Standard Error}
\usage{
coverage_MCMC_SE(coverage, nsim)
}
\arguments{
\item{coverage}{scalar. Coverage of the scenario under investigation.}

\item{nsim}{scalar. Number of iterations used in the scenario.}
}
\value{
MCMC standard error of coverage.
}
\description{
Coverage MCMC Standard Error
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
