% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{empirical_MCMC_SE}
\alias{empirical_MCMC_SE}
\title{Compute the MCMC SE}
\usage{
empirical_MCMC_SE(emp_se, nsim)
}
\arguments{
\item{x}{numeric vector. Simulated parameter estimates.}
}
\value{
numeric scalar. MCMC standard error.
}
\description{
Compute the MCMC SE
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
