% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{empirical_SE}
\alias{empirical_SE}
\title{Compute the empirical SE}
\usage{
empirical_SE(x)
}
\arguments{
\item{x}{numeric vector. Simulated parameter estimates}
}
\value{
numeric vector with two values: (1) empirical standard error and (2) MCMC SE of this value
}
\description{
Compute the empirical SE
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
