% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jasper.R
\name{get_cci}
\alias{get_cci}
\title{Retrieve CCI estimates for parameters}
\usage{
get_cci(x)
}
\arguments{
\item{x}{an mHMM object}
}
\value{
List. Confidence credibility intervals (CCIs) estimates for each parameter.
        names of the elements are identical of the names of the input
        parameters
}
\description{
Retrieve CCI estimates for parameters
}
