% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jasper.R
\name{get_subject_tpm}
\alias{get_subject_tpm}
\title{Convert a list of probabilities into an m x m transition probability matrix}
\usage{
get_subject_tpm(x, ...)
}
\arguments{
\item{x}{an mHMM object}
}
\value{
m x m transition probability matrix, where m is the number of hidden states
}
\description{
Convert a list of probabilities into an m x m transition probability matrix
}
