% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_sim.R
\name{init_log}
\alias{init_log}
\title{Initialize simulation log}
\usage{
init_log(
  sim_sample_size = 30,
  sim_n_t = c(400, 800, 1600),
  sim_m = 3,
  sim_n_dep = c(1, 2, 4),
  sim_q_emiss = 5,
  sim_gamma_var = 1,
  sim_emiss_var = 1,
  sim_noisiness = c(0.03, 0.09, 0.15),
  sim_overlapping = c("low", "moderate", "high"),
  sim_iter = 5000,
  sim_burnin = 1750,
  sim_repetitions = 1:250,
  save_frac_res = 0.02,
  seed = 123
)
}
\description{
Initialize simulation log
}
