% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jasper.R
\name{MAP}
\alias{MAP}
\title{Retrieve MAP estimates for parameters}
\usage{
MAP(x, ...)
}
\arguments{
\item{x}{an mHMM object}
}
\value{
List. Maximum a Posteriori (MAP) estimates for each parameter.
        names of the elements are identical of the names of the input
        parameters
}
\description{
Retrieve MAP estimates for parameters
}
