% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMMfast.R
\name{mHMMfast}
\alias{mHMMfast}
\title{Get estimations with mHMMfast}
\usage{
mHMMfast(
  s_data,
  gen,
  xx = NULL,
  start_val,
  mcmc,
  return_path = FALSE,
  print_iter,
  show_progress = TRUE,
  gamma_hyp_prior = NULL,
  emiss_hyp_prior = NULL,
  gamma_sampler = NULL,
  emiss_sampler = NULL
)
}
\value{
#' @return \code{mHMM} returns an object of class \code{mHMM}, which has
  \code{print} and \code{summary} methods to see the results.
}
\description{
Get estimations with mHMMfast
}
