% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHMMlight.R
\name{mHMMlight}
\alias{mHMMlight}
\title{Get estimations with mHMMlight}
\usage{
mHMMlight(
  s_data,
  gen,
  xx = NULL,
  start_val,
  mcmc,
  return_path = FALSE,
  print_iter,
  show_progress = TRUE,
  gamma_hyp_prior = NULL,
  emiss_hyp_prior = NULL,
  gamma_sampler = NULL,
  emiss_sampler = NULL,
  save_subj_data = FALSE
)
}
\value{
#' @return \code{mHMM} returns an object of class \code{mHMM}, which has
  \code{print} and \code{summary} methods to see the results.
}
\description{
Get estimations with mHMMlight
}
