% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_results.R
\name{postprocess_ci}
\alias{postprocess_ci}
\title{Postprocess credible intervals}
\usage{
postprocess_ci(z, m)
}
\arguments{
\item{z}{list. contains the credible intervals of the transition probabilities.}

\item{m}{integer. Number of hidden states}
}
\value{
data frame in which each CCI has been placed in its own column. (with _lower and _upper appended).
}
\description{
Postprocess credible intervals
}
