% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_results.R
\name{postprocess_subject_specific}
\alias{postprocess_subject_specific}
\title{Postprocess subject-specific metrics}
\usage{
postprocess_subject_specific(z, m)
}
\arguments{
\item{z}{list. contains the MAP values of the parameter estimates for each subject.}

\item{m}{integer. Number of hidden states}
}
\value{
data frame with rows equal to the number of subjects and columns equal to the number of states x dependent variables x
}
\description{
Postprocess subject-specific metrics
}
