% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{rel_bias}
\alias{rel_bias}
\title{Compute the parameter bias of a vector estimated parameters versus the ground-truth value of that parameter (modified from Jasper)}
\usage{
rel_bias(true_param_value, simulated_param_values)
}
\arguments{
\item{true_param_value}{numeric. Value of the ground-truth parameter.}

\item{simulated_param_values}{k-length numeric vector. k >= 1 and holds the parameter values of the estimated parameters.}
}
\value{
numeric vector. Contains two elements: (1) average bias of simulated values versus the ground-truth value and (2) MCMC SE of the bias value
}
\description{
Compute the parameter bias of a vector estimated parameters versus the ground-truth value of that parameter (modified from Jasper)
}
\details{
This function computes the percentage bias by using the signed mean difference.
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
