% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{rel_bias_MCMC_SE}
\alias{rel_bias_MCMC_SE}
\title{MCMC Standard Error of the bias value}
\usage{
rel_bias_MCMC_SE(x, y)
}
\arguments{
\item{x}{k-length numeric vector. k >= 1 and holds the parameter values of the estimated parameters.}
}
\value{
Bias MCMC SE
}
\description{
MCMC Standard Error of the bias value
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
