% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_sim_surf.R
\name{run_one_sim_surf}
\alias{run_one_sim_surf}
\title{Run one simulation in SurfSara Lisa}
\usage{
run_one_sim_surf(pars, light = FALSE, save_subj_data = TRUE, baseline = FALSE)
}
\description{
Run one simulation in SurfSara Lisa
}
