% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{summarize_simulation_scenario}
\alias{summarize_simulation_scenario}
\title{Wrapper function that computes all simulation metrics}
\usage{
summarize_simulation_scenario(
  true_values,
  simulated_values,
  lower_cci,
  upper_cci,
  SE = NULL,
  compute_multimodal = FALSE
)
}
\arguments{
\item{true_values}{vector of length n. Contains true population values.}

\item{simulated_values}{vector of length n. Contains simulated values.}

\item{lower_cci}{vector of length n. Upper 95\% CCI.}

\item{SE}{vector of length n. Standard Error of the posterior distribution.}

\item{compute_multimodal}{boolean. If TRUE, this function will compute a test on the simulated parameter values to check if the distribution of parameter estimates is multimodal. See \link[multimode]{modetest}.}
}
\value{
data frame containing:
\describe{
  \item{bias}{percent bias of the simulation scenario, computed as a percentage relative to the true value.}
  \item{bias_mcmc_se}{MCMC standard error of bias estimate, computed as a percentage relative to the bias estimate.}
  \item{empirical_se}{empirical standard error computed from the simulated values.}
  \item{empirical_se_mcmc_se}{MCMC standard error of the empirical SE.}
  \item{modSE}{model standard error computed from the simulated values.}
  \item{modSE_mcmc_se}{MCMC standard error of the model SE.}
  \item{MSE}{Mean-Squared Error of the simulated values.}
  \item{MSE_mcmc_se}{MCMC standard error of the simulated values.}
  \item{coverage}{Coverage given the 95\% CCI.}
  \item{coverage_mcmc_se}{MCMC standard error of coverage.}
  \item{bias_corr_coverage}{Bias-adjusted coverage. Instead of using the true population value, use the mean of the simulated values. Useful to check whether poor coverage is the result of bias. See 'See also' for reference.}
  \item{bias_corr_coverage_mcmc_se}{MCMC standard error of bias-adjusted coverage.}
  \item{multimodal}{p-value of the \link[multimode]{modetest} used to check for multimodal distributions.}
}
}
\description{
This wrapper function is compatible with dplyr and can be called in a dplyr chain
 of arguments.
}
\examples{
\dontrun{
tst <- data \%>\%
   group_by(scenario_id) \%>\%
   # This is how you use the function
   do(summarize_simulation_metrics(.$true_values, .$simulated_values
                                   .$lower_cci, .$upper_cci, FALSE))
}

}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
