% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.r
\name{test_fit_mHMM}
\alias{test_fit_mHMM}
\title{Fit mHMM using mHMMfast}
\usage{
test_fit_mHMM(
  m,
  n_dep,
  q_emiss,
  gamma,
  emiss,
  iter,
  burnin,
  start_gamma = NULL,
  start_emiss = NULL,
  data_sim,
  light = FALSE,
  save_path = FALSE,
  save_subj_data = TRUE,
  progress_line = FALSE,
  good_bad = "good"
)
}
\description{
Fit mHMM using mHMMfast
}
