% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.r
\name{test_run_one_sim_surf}
\alias{test_run_one_sim_surf}
\title{Run one simulation in SurfSara Lisa}
\usage{
test_run_one_sim_surf(
  pars,
  light = FALSE,
  save_subj_data = TRUE,
  baseline = FALSE,
  convergence = FALSE,
  save_path = FALSE,
  progress_line = FALSE
)
}
\description{
Run one simulation in SurfSara Lisa
}
