% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vit_mHMM_group_map.R
\name{vit_mHMM_group_map}
\alias{vit_mHMM_group_map}
\title{Obtain hidden state sequence for each subject using the Viterbi
algorithm}
\usage{
vit_mHMM_group_map(object, s_data, data_distr)
}
\description{
Obtain hidden state sequence for each subject using the Viterbi
algorithm
}
